# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KIOSK API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/hod/overview": {
      "get": {
        "tags": [
          "hod"
        ],
        "summary": "Overview",
        "operationId": "overview_api_v1_hod_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/hod/counsellors": {
      "get": {
        "tags": [
          "hod"
        ],
        "summary": "Counsellors",
        "operationId": "counsellors_api_v1_hod_counsellors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/hod/students": {
      "get": {
        "tags": [
          "hod"
        ],
        "summary": "Students",
        "operationId": "students_api_v1_hod_students_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/hod/assign": {
      "post": {
        "tags": [
          "hod"
        ],
        "summary": "Assign",
        "operationId": "assign_api_v1_hod_assign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/hod/assign-section": {
      "post": {
        "tags": [
          "hod"
        ],
        "summary": "Assign Section",
        "operationId": "assign_section_api_v1_hod_assign_section_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSectionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/hod/assign/{register_no}": {
      "delete": {
        "tags": [
          "hod"
        ],
        "summary": "Unassign",
        "operationId": "unassign_api_v1_hod_assign__register_no__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "register_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Register No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hod/reports/leave.csv": {
      "get": {
        "tags": [
          "hod"
        ],
        "summary": "Leave Report",
        "operationId": "leave_report_api_v1_hod_reports_leave_csv_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeaveStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hod/reports/attendance-shortage.csv": {
      "get": {
        "tags": [
          "hod"
        ],
        "summary": "Shortage Report",
        "operationId": "shortage_report_api_v1_hod_reports_attendance_shortage_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/hod/students/import": {
      "post": {
        "tags": [
          "hod"
        ],
        "summary": "Import Sheet",
        "description": "Upload the college's student Excel sheet (.xlsx). Check first with dryRun=true, then send again with\ndryRun=false. Only rows of the HOD's own department are accepted. Slow for big sheets (passwords are hashed).",
        "operationId": "import_sheet_api_v1_hod_students_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_sheet_api_v1_hod_students_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignIn": {
        "properties": {
          "counsellorId": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Counsellorid"
          },
          "registerNos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Registernos"
          }
        },
        "type": "object",
        "required": [
          "counsellorId",
          "registerNos"
        ],
        "title": "AssignIn"
      },
      "AssignSectionIn": {
        "properties": {
          "counsellorId": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Counsellorid"
          },
          "section": {
            "type": "string",
            "maxLength": 5,
            "minLength": 1,
            "title": "Section"
          }
        },
        "type": "object",
        "required": [
          "counsellorId",
          "section"
        ],
        "title": "AssignSectionIn"
      },
...
      "Body_import_sheet_api_v1_hod_students_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "dryRun": {
            "type": "boolean",
            "title": "Dryrun",
            "default": true
          },
          "assignTo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignto"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_sheet_api_v1_hod_students_import_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LeaveStatus": {
        "type": "string",
        "enum": [
          "PENDING_COUNSELLOR",
          "REJECTED_BY_COUNSELLOR",
          "PENDING_HOD",
          "APPROVED",
          "REJECTED_BY_HOD"
        ],
        "title": "LeaveStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}